Seems to be a problem when Install not in target ns
Debugging this made me realize we need to filter the resources prior
to checking the Install CR's version. Otherwises, the namespaces may
be wrong, which can happen when the operator is started with different
options. This can of course cause the operator to be wrong about the
current install state of the operand.