Uses 'lessolm' branch of manifestival to fix various OLM issues
Turns out we don't need to prevent the operator from installing
the manifest resources typically installed by OLM. We just need to
make sure we set the namespace references correctly and give the
operator enough permissions to create them.

For now, this is cluster-admin, but we can revisit that.

One huge benefit is that we don't need to extract any CRD's or
replicate the upstream RBAC to that of the operator. This should
greatly simplify subsequent releases.