Automatically install the Maistra Operator and ControlPlane
On OpenShift, we detect if the Maistra Operator CRD is already
installed. If not, and Istio CRDs are also not already present, then
we install the Maistra Operator and a minimal Maistra ControlPlane.

If we detect that the Maistra Operator is already installed but no
Maistra ControlPlane then we install the same minimal Maistra
ControlPlane.