Ensure we have a single KnativeServing resource [SRVKS-129]
The operator will now stick a KnativeServing CR in knative-serving
when it starts up, so installation is immediate.

We monitor all namespaces for CR's now, so...
  $ operator-sdk up local --namespace=""

We moved the manifest creation to InjectClient to clean up some of the
init logic.

We addressed a bug in manifestival that resulted in optimisitic
locking errors.