Remove short-circuiting installation logic if we're in an installing state.
This piece of code is intended to prevent us from reapplying the full YAML if we're already in a deploying state. It can cause us to not recreate deployments though if we end up in an interim state, i.e. we're already in a deploying state and deployments start to vanish. We will never reconcile that in that case because this piece of code will always say that we're deploying and the checkDeployments code will never tell us something different.

Reapplying the YAML in any case is benign anyway. We can be smarter about this later and check the spec of what's already applied against the spec of the YAML in the file. That's purely a performance optimization though, correctness should not be affected by that.