Introduce immutable transforms and consider KnativeServing CR config to be canonical (#165)
* Add an e2e test for #138

After creating the KnativeServing instance, we then update it twice,
once with an entry for the config-logging ConfigMap and then again
without the key, verifying that the operator correctly syncs the
ConfigMap each time.

* Uses a pre-release of manifestival to fix #138

This relies on the Transform function now being immutable, returning a
new Manifest instead of mutating its source resources.

This is related to #163 as well.

* Address golint warnings

* Transform manifest once and pass to each reconcile stage

* Let's confine the arbitrariness of the configmap to the fixture

This also makes the WaitForConfigMap helper more helpful