Update to Manifestival v0.1.0 (#297)
* Update to latest manifestival

This introduces support for different client implementations, e.g.
client-go and controller-runtime. This repo uses client-go so that
requires a dependency on the repo implementing the client-go
manifestival client.

* Bump to first released version of Manifestival, v0.1.0

* Don't expose recursive manifest creation to user

This is potentially dangerous when the user has no control over the
directory being searched. And currently, we have no sub-directories
beneath kodata/knative-serving/ anyway.