Run the upgrade and downgrade tests in the upgrade prow job (#300)
* Change the test namespace from operator-tests to knative-serving

We change the test namespace, because all the tests in serving repository
are running under the namespace knative-serving. We will call the upgrade
and downgrade tests directly in serving. We need to make sure all the tests
are running under the same namespace.

* Add the script upload-test-images.sh to build and publish the test images

* Change the postupgrade test for operator

* We do not need to delete the operator CR, becuase we will verify the
services in serving.
* We do not need to create the opertor CR, because we will install a
previous version of the operator and create CR before reaching this
test.
* We need to verify the status of all the deployments, making sure
they match the expected deployments in terms of number and names,
and reach the ready status.

* Refactor the e2e-common.sh script for adding upgrade and downgrade tests

* Run the upgrade and downgrade tests in serving