Delete roles and rolebindings as the final step of cleanup.
This allows us to take advantage of the permissions granted to cluster admins
when performing cleanup. If the approach in https://github.com/knative/serving-operator/pull/291
is approved, this will become necessary. Additionally, it makes sense to remove roles as the final
step to allow human Operators to modify any resources they may have permissions on as a result of the
Knative installation (that is, we should not remove any access until we are 'almost done' cleaning up).