Resources override (#362)
* New field in CR, `resources`, to override deployments in manifest

It's a map of deployment names to a []v1.ResourceRequirement. The
array corresponds to the order of the containers in the deployment
spec.

* Use 0.13.0 manifest for some tests

* Unit tests for resource overrides

Added a dummy container to the webhook just to show that capability,
even if yet unused

* Change-of-plan: key by container name instead of deployment name

This removes any assumption of container order in the deploy spec,
with the minor trade-off of container names expected to be unique
within the manifest.

* Son of change-of-plan: no more default removal of resource reqs

We now merge the resource overrides with the existing specs, so that
the upstream defaults remain intact if there are no overrides in the
KnativeServing CR.

We also converted the map to an array to abide by k8s conventions.