Respect HA Replicas setting in HPAs.
If the operator has a HighAvailabiliy configuration and the replicas are set > 1, we probably want to adjust all HPAs to have at least as many minReplicas as specified in that HA config. An HPA being present is enough of a signal for us to assume that the component supports a HA configuration.